perf(perms): 用户/用户组授权的所有应用API返回attrs属性
