+++ apps/perms/api/application/user_group_permission.py
@@ -6,6 +6,7 @@ from rest_framework.generics import ListAPIView
 
 from common.permissions import IsOrgAdminOrAppUser
 from applications.models import Application
+from applications.api.mixin import ApplicationAttrsSerializerViewMixin
 from perms import serializers
 
 __all__ = [
@@ -13,14 +14,14 @@ __all__ = [
 ]
 
 
-class UserGroupGrantedApplicationsApi(ListAPIView):
+class UserGroupGrantedApplicationsApi(ApplicationAttrsSerializerViewMixin, ListAPIView):
     """
     获取用户组直接授权的资产
     """
     permission_classes = (IsOrgAdminOrAppUser,)
     serializer_class = serializers.ApplicationGrantedSerializer
     only_fields = serializers.ApplicationGrantedSerializer.Meta.only_fields
-    filter_fields = ['id', 'name', 'comment']
+    filter_fields = ['id', 'name', 'category', 'type', 'comment']
     search_fields = ['name', 'comment']
 
     def get_queryset(self):
